Now playable in Chrome :)
The SimpleBrick doesn't work because JavaScript doesn't have integer
types, and therefore does not do integer division (the fancy bit tricks
for keeping track of bricks is therefore broken).

Also, the canvas does some sort of anti-aliasing on lines that are a
single pixel wide, so for now all lines are drawn two pixels wide (which
breaks some of the nice graphics).