Now fully functional in Chrome as well!
Fixed bug with anti-aliasing when using stroke rectangles.
The missing JavaScript integer-division was fixed with some sneaky hacks
(that will break if the ThingML model is changed).

There seems to be a bug in the State.js library when messages that trigger
transitions - are sent in transition actions. For the SimpleBrick code
this was fixed by changing a self-transition to an internal event.